+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,7 @@
 #### 🐞 Fixes
 
 - Fixed an issue with codegen templates not working with asset/binary files correctly.
+- Fixed an issue where `moon upgrade` would require a workspace.
 
 ## 1.13.0