+++ packages/cli/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 #### 🐞 Fixes
 
+- Fixed an issue where tasks depending on arbitrary project tasks, would fail to build a partial
+  project graph.
 - Fixed an issue with asset/binary files not working correctly with codegen templates.
 - Fixed an issue where `moon upgrade` would require a workspace.